+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 - 更换背景图片功能支持设置无背景图片 | The background image replacement function supports setting no background image.
 - 优化扩展菜单的分割线 | Optimize the separator of the extension menu.
+- 禁用浏览器中 <kbd>Ctrl/⌘ + 鼠标滚轮</kbd> 缩放 DOM | Disable the DOM zooming with <kbd>Ctrl/⌘ + Mouse Wheel</kbd> in the browser.
 
 ## v1.8.0/2023-04-12